requirements.txt
# Core & Data Handling
neo4j
networkx
numpy
scikit-learn

# NLP & Machine Learning
torch
transformers
sentence-transformers
spacy
datasets
nltk
sumy

# Web & APIs
requests
beautifulsoup4
google-api-python-client
google-generativeai
aiohttp

# Voice Interaction
pyttsx3
SpeechRecognition

# Environment
gymnasium

# Utility
dataclasses